Refactor UniquenessCheckResponse to ledger
