+++ ledger/ledger-utxo/build.gradle
@@ -14,6 +14,7 @@ dependencies {
     api project(':application')
     api project(':base')
     api project(':crypto')
+    api project(':ledger:ledger-common')
 
     compileOnly 'org.osgi:osgi.core'
     compileOnly 'org.osgi:osgi.annotation'
