+++ ledger/ledger-utxo/src/main/kotlin/net/corda/v5/ledger/utxo/uniqueness/client/LedgerUniquenessCheckResponse.kt
@@ -1,6 +1,7 @@
-package net.corda.v5.application.uniqueness.model
+package net.corda.v5.ledger.utxo.uniqueness.client
 
-import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
+import net.corda.v5.application.uniqueness.model.UniquenessCheckResult
+import net.corda.v5.ledger.common.transaction.TransactionSignature
 
 /**
  * This is a response class that wraps the result of the uniqueness checking and a potential signature,
@@ -10,7 +11,7 @@ import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
  * @property signature Contains the signature provided by the uniqueness checker. Will be `null` if the
  * result is a failure.
  */
-interface UniquenessCheckResponse {
+interface LedgerUniquenessCheckResponse {
     val result: UniquenessCheckResult
-    val signature: DigitalSignatureAndMetadata?
+    val signature: TransactionSignature?
 }
